Add a way to map package names from one distro to another
To support different distributions, we currently list packages needed
for each distribution. This is not fun, especially as most package names
are common. Also, to properly support this in all the code, we need to
add several if/elif/else/fi.

Instead, we can use a map to match the package names from one distro
(that we'll consider as a basis, with package names used in the code) to
the package names of another distribution. This means we simply need to
add a file containing the mapping to support this other distribution --
at least, in terms of packaging.

Mapping files are in files/package-maps/ and are named after the
vendors. Format is the following:

 a) To map a single package to one or more packages:
      original_name first_package,second_package
 b) To keep the same name for the mapping, just put:
      original_name
    (or alternatively, do not even put the file in the mapping, but we
    might add tests to make sure there's a mapping defined for each
    package)
 c) To accomodate a package not existing in the target distro, use:
      original_name -
    This will simply skip this package for the target distro.

Change-Id: I01b29ec73a7294f11d74adcf2c9a62797d88383c